Added seperate buttons for debit and credit transactions of bank and cash - Reason for changing the UI is in future I am planning to add other fields while adding transaction like categories, from which bank account money debited, credit to which bank account etc, so for that the current add transaction form doesn't have much space - After following above UI I will not have to show type and mode of transaction choice so that space will be removed - Also its more intutive to select the type and mode of transaction just by a button click instead first opening the form and then selecting the type or mode of transaction - For now just looking how the UI will look at mobile later in next commit will add the logic
Refactored the code to use typescript and scss
- Will later follow BEM convention of naming
