+++ package.json
@@ -28,6 +28,7 @@
     "redux-devtools": "^3.7.0",
     "redux-devtools-extension": "^2.13.8",
     "redux-persist": "^6.0.0",
+    "sass": "^1.32.8",
     "styled-components": "^5.2.1",
     "typescript": "^4.1.3",
     "web-vitals": "^0.2.4",
