+++ src/components/TransactionSummaryAndAdd/styles.scss
@@ -1,0 +1,34 @@
+.transaction-summary {
+  margin: 5px 5px;
+  padding: 5px;
+  border-radius: 5px;
+  box-shadow: 0 1px lightgrey;
+  font-size: 18px;
+  font-weight: bold;
+
+}
+
+.transaction-summary-heading {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 10px;
+  padding: 0 10px;
+}
+
+.credit-debit-button {
+  display: inline-block;
+  padding: 8px;
+  margin: 0 10px;
+  border: none;
+  outline: none;
+  border-radius: 5px;
+}
+
+.credit {
+  background-color: lightgray;
+}
+
+.debit {
+  background-color: lightgray;
+}
