+++ src/components/Transactions/styles.scss
@@ -1,0 +1,4 @@
+.transactions-list {
+  margin: 0 5px;
+  padding: 0;
+}
